% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findre.R
\name{buildfas}
\alias{buildfas}
\title{Build FASTA file}
\usage{
buildfas(name, sequences, file_name)
}
\arguments{
\item{name}{String, the name of the sequence}

\item{sequences}{String, the sequences}

\item{file_name}{String, The name of FASTA file}
}
\value{
Create a FASTA file and return the it's name
}
\description{
A function helps users to build their own FASTA file
}
\examples{
seq <- "ACGTCG"
name <- "Target"
file <- buildfas(name, seq, 'tempfasta.fas')
fas_seq <- seqinr::read.fasta(file, as.string = TRUE)
attr(fas_seq, "name")
toString(fas_seq)

}
