% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmap.R
\name{check_dou}
\alias{check_dou}
\title{Helper function to check if the fragments satisify the double digests reaction}
\usage{
check_dou(frag1, frag2, dou_dig)
}
\arguments{
\item{frag1}{the int vector of result of
the single digests reaction by enzyme1.}

\item{frag2}{the int vector of result of
the single digests reaction by enzyme2.}

\item{dou_dig}{the int vector of result of the
double digests reaction by enzyme1 and enzyme2.}
}
\value{
if the fragments cut by enz1 and enz2 is
  same with double digests reaction
}
\description{
The helper function for \code{rsmap} to check
if the fragments satisify the double digests reaction
}
