% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cutpos}
\alias{cutpos}
\title{Find the position cut by given enzyme}
\usage{
cutpos(dnaseq, enz, dataset = redata)
}
\arguments{
\item{dnaseq}{a string of DNA sequence}

\item{enz}{the name of enzyme}

\item{dataset}{the data used to store enzyme info (name and site)}
}
\value{
a data frame of cut position (name, seq, start and end), if not found, return NA.
}
\description{
A function to locate the cut site of given enzyme and enzyme dataset.
}
\examples{
dnaSeq <- 'ATCGGTTATAAGCAT'
enz <- 'AanI'
cutpos(dnaSeq, enz, redata)

}
