% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redata}
\alias{redata}
\title{777 Restriction Enzymes and recognition sites.}
\format{A data frame with 777 rows and 2 variables:
\describe{
  \item{name}{name of the restriction enzyme}
  \item{site}{the recognition sites of the restriction enzymes}
  ...
}}
\source{
\url{http://rebase.neb.com}
}
\usage{
redata
}
\description{
A dataset containing the 777 Restriction Enzymes
}
\keyword{datasets}
