% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmap.R
\name{rsmap}
\alias{rsmap}
\title{Construct restriction map by given fragments}
\usage{
rsmap(enz1, frag1, enz2, frag2, dou_dig, name = "Unknown Sequence")
}
\arguments{
\item{enz1}{the name of first enzyme.}

\item{frag1}{the int vector of result of
the single digests reaction by enzyme1.}

\item{enz2}{the name of second enzyme.}

\item{frag2}{the int vector of result of
the single digests reaction by enzyme2.}

\item{dou_dig}{the int vector of result of the
double digests reaction by enzyme1 and enzyme2.}

\item{name}{the name of tested sequence}
}
\value{
restriction map
}
\description{
This function will visualize the restriction site map,
by given single and double digests reaction result.
Assume the input sequence is circular.
}
\examples{
frag1 <- c(100)
frag2 <- c(50, 25, 20, 5)
dou_dig <- c(25, 25, 25, 20, 5)
enz1 <- 'enz1'
enz2 <- 'enz2'
result <- rsmap(enz1, frag1, enz2, frag2, dou_dig)
result

}
