% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsplot.R
\name{rsplot}
\alias{rsplot}
\title{plot restriction map}
\usage{
rsplot(dna, target, dataset, num)
}
\arguments{
\item{dna}{The FASTA file of the sequence}

\item{target}{The FASTA file of the target sequence you want to produce}

\item{dataset}{The dataset stores the information of enzymes}

\item{num}{The number of enzyme on the left/right side of target sequence}
}
\description{
plot restriction map
}
