% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRsrmApp.R
\name{runRsrmApp}
\alias{runRsrmApp}
\title{Launch the shiny app for package rsrm}
\usage{
runRsrmApp()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runRsrmApp()
}

}
