% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findre.R
\name{buildfas}
\alias{buildfas}
\title{Build FASTA file}
\usage{
buildfas(name, sequences, file_name)
}
\arguments{
\item{name}{The name of the sequence}

\item{sequences}{The string of sequences}

\item{file_name}{The name of FASTA file}
}
\value{
a FASTA file
}
\description{
A function helps users to build their own FASTA file
}
