% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cutpos}
\alias{cutpos}
\title{Find the position cut by given enzyme}
\usage{
cutpos(dnaSeq, enz, dataset = redata)
}
\arguments{
\item{dnaSeq}{a string of DNA sequence}

\item{enz}{the name of enzyme}

\item{dataset}{the data used to store enzyme info}
}
\value{
a matrix of cut position
}
\description{
Find the position cut by given enzyme
}
