% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findre.R
\name{findre}
\alias{findre}
\title{Find the Restriction Sites on given DNA sequence}
\usage{
findre(dnaName, dnaSeq, targetName = "Target", targetSeq,
  dataset = redata, num = 6,
  title = "Restriction map around the target sequence")
}
\arguments{
\item{dnaName}{String, The name of the DNA sequence,}

\item{dnaSeq}{String, The sequence of the DNA}

\item{targetName}{String, The name of the target you want to produce, defualt is just "TARGET".}

\item{targetSeq}{String, The sequence of the target you want to produce.}

\item{dataset}{The dataset stores the information of enzymes with two column, name and site.}

\item{num}{The number of enzyme on the left/right side of target sequence, defualt is 6}

\item{title}{The title of the plot.defualt is "Restriction map around the target sequence"}
}
\value{
the distribution of restriction site arround the target sequence
}
\description{
A function that find Restriction Sites on given DNA sequence by given restriction enzyme dataset.
and plot the restriction site map by given number of sites you can choose how many sites at left/right of target to display.
}
\examples{
seq1 <- 'GGCAGATTCCCCCTAACGTCGGACCCGCCCGCACCATGGTCAGGCATGCCCCTCCTCATCGCTGGGCACAGCCCAGAGGGT
ATAAACAGTGCTGGAGGCTGGCGGGGCAGGCCAGCTGAGTCCTGAGCAGCAGCCCAGCGCAGCCACCGAGACACC
ATGAGAGCCCTCACACTCCTCGCCCTATTGGCCCTGGCCGCACTTTGCATCGCTGGCCAGGCAGGTGAGTGCCCC'
name1 <- 'example DNA Sequence'
seq2 <- 'ACGTCG'
name2 <- 'Target'
result <- findre(name1, seq1, name2, seq2)
result

}
