% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findre.R
\name{findre}
\alias{findre}
\title{Find the Restriction enzymes which can produce the target sequence}
\usage{
findre(dna, target, dataset = redata)
}
\arguments{
\item{dna}{The FASTA file of the sequence}

\item{target}{The FASTA file of the target sequence you want to produce}

\item{dataset}{The dataset stores the information of enzymes}
}
\value{
a list of Restriction enzymes combination
}
\description{
A function that find a combination of the restriction enzymes
which can produce a fragment contains target sequnce by digest reaction
}
