% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rs2regex}
\alias{rs2regex}
\title{convert recognition site to regular expression}
\usage{
rs2regex(rs)
}
\arguments{
\item{rs}{a string of recognition site}
}
\value{
a string of regex
}
\description{
convert recognition site to regular expression
}
