% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sanitizeSeq}
\alias{sanitizeSeq}
\title{Remove invalid characters in DNA sequence}
\usage{
sanitizeSeq(s)
}
\arguments{
\item{s}{chr DNA sequence, maybe not vaild}
}
\value{
chr a valid, uppercase, DNA sequence
}
\description{
Remove FASTA header and all letters except A, T, C, G.
}
